# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "https://archimed.justcode.am/",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/medical_test": {
            "post": {
                "tags": [
                    "MedicalTest"
                ],
                "summary": "medical_test",
                "description": "Получаем Список Комплексов анализа",
                "operationId": "MedicalTest",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "search": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "search"
                                    },
                                    "complex": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Test Data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/medical_test_params": {
            "post": {
                "tags": [
                    "MedicalTest"
                ],
                "summary": "medical_test_params",
                "description": "Получаем Список Анализов  конкретного комплекса",
                "operationId": "medical_test_params",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "search": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "search"
                                    },
                                    "test_id": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "39"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Test Analize Data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/medical_test_params_single_page": {
            "post": {
                "tags": [
                    "MedicalTest"
                ],
                "summary": "Get medical test parameters for a single page",
                "operationId": "4ddf6e2d52e11995633de80381730e38",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "test_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong test_id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Wrong test_id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "LABEL": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}